[*] fail при загрузке скрипта с пробелами в названии [*] AV в окне настроек

+++ branches/1.6.1/Forms/frm_main.pas
@@ -4152,35 +4152,34 @@ begin
 
       if (Data.nodeType = ntBookInfo) and ((Tree.CheckState[Node] = csCheckedNormal) or (Tree.Selected[Node])) then
       begin
-        if (isOnline and Data.Locale)
-           or Settings.DeleteFiles
-         then
+        dmCollection.GetBookFileName(Data.ID, FileName, Folder, Ext, No);
+        if (isOnline and Data.Locale) then
         begin
-          dmCollection.GetBookFileName(Data.ID, FileName, Folder, Ext, No);
-
-          if not isFB2 then
-          begin
-            if (ExtractFileExt(FileName) = ZIP_EXTENSION) then
-              DeleteFile(FCollectionRoot + Folder + FileName)
-            else
-              DeleteFile(FCollectionRoot + Folder + FileName + Ext);
-          end;
-
-          if isFB2 and isPrivate
-          then
-          begin
-            if (ExtractFileExt(Folder) = ZIP_EXTENSION) then
-              DeleteFile(FCollectionRoot + Folder)
-            else
-              DeleteFile(FCollectionRoot + Folder + FileName + Ext);
-          end;
-
           if DeleteFile(FCollectionRoot + Folder) and isOnline
           then
             dmCollection.SetLocalStatus(Data.ID,False);
         end
         else
         begin
+          if  Settings.DeleteFiles then
+          begin
+            if not isFB2 then
+            begin
+              if (ExtractFileExt(FileName) = ZIP_EXTENSION) then
+                DeleteFile(FCollectionRoot + Folder + FileName)
+              else
+                DeleteFile(FCollectionRoot + Folder + FileName + Ext);
+            end;
+
+            if isFB2 and isPrivate
+            then
+            begin
+              if (ExtractFileExt(Folder) = ZIP_EXTENSION) then
+                DeleteFile(FCollectionRoot + Folder)
+              else
+                DeleteFile(FCollectionRoot + Folder + FileName + Ext);
+            end;
+          end;
           ALibrary.BeginBulkOperation;
           try
             ALibrary.DeleteBook(Data.ID);
